# bpo-46420: Use NOTRACE_DISPATCH() in specialized opcodes