@markshannon 2022-01-18
It is good to see that removing many tracing checks can produce a measurable speedup.
Unfortunately, I don't think this is correct. It is possible for tracing to be turned on during several of these bytecodes.

Most obviously, `CALL_NO_KW_BUILTIN_CLASS_1` and `CALL_NO_KW_BUILTIN_FAST` could end up calling `sys.settrace`.
More subtly, any allocation can trigger the cycle collector, which can execute arbitrary code.

The approach I want to take, although it needs more work, is to only check for tracing in the `RESUME` instruction by using instrumentation instead.

See https://github.com/faster-cpython/ideas/issues/112 and [PEP 669](https://www.python.org/dev/peps/pep-0669/).
Hopefully, using instrumentation should provide at least the speedup you have observed, *and* reduce the performance loss when tracing is used.