@sweeneyde 2022-01-18
Since `STORE_FAST__STORE_FAST` calls `SETLOCAL`, which can call `Py_DECREF`, which can execute arbitrary `__del__` methods, does that mean that existing usage of `NOTRACE_DISPATCH` is mildly incorrect as well?