@markshannon 2022-01-19
Perhaps.
I think it is OK, because the `__del__` is not specified to occur at any specific time, or in any particular order.
So, the behavior is the same *as* *if* we checked for tracing on every dispatch.

Most of your changes are also probably safe. Probably. It is hard to be sure.
Although I do think the changes to `CALL` specializations are incorrect.

The instrumentation approach is much more robust. At least, I think it is.