@markshannon 2022-01-25
Looks good. All the new uses of `NOTRACE_DISPATCH()` are consistent with existing usage.